Update README.md: Added command for building docker image on Mac
AWS Fargate required a Docker image for linux/amd64 platform.

If the Docker image is built on Macs with Apple Silicon, it is built for the arm architecture, which causes failure after deployment.

Hence, added the command to build the image for the linux/amd64. 

Please merge these changes, so Mac users won't face issues in building this project.